huge commit: changes all tests that use a db file to execute against all supported db file formats. The list of formats to test is easily changed via JetFormatTest.SUPPORTED_FILEFORMATS.
All tests pass with v2000.
One test fails with v2003.
A half dozen fail with v2007 (but many are the same read error).

includes re-org of test db files for easier support of multiple file formats.

(this commit is for ONLY the test db files - source changes are in prior commit).

git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/branches/newformats@442 f203690c-595d-4dc9-a70b-905162fa7fd2